Move `FeedImageDataLoaderWithFallbackComposite` to production
