+++ EssentialApp/EssentialApp/FeedImageDataLoaderWithFallbackComposite.swift
@@ -1,0 +1,41 @@
+//
+//  FeedImageDataLoaderWithFallbackComposite.swift
+//  EssentialApp
+//
+//  Created by Viral on 06/10/22.
+//
+
+import Foundation
+import EssentialFeed
+
+public class FeedImageDataLoaderWithFallbackComposite: FeedImageDataLoader {
+    let primaryLoader: FeedImageDataLoader
+    let fallbackLoader: FeedImageDataLoader
+
+    private class TaskWrapper: FeedImageDataLoaderTask {
+        var wrapper: FeedImageDataLoaderTask?
+        func cancel() {
+            wrapper?.cancel()
+        }
+    }
+
+    public init(primaryLoader: FeedImageDataLoader, fallbackLoader: FeedImageDataLoader) {
+        self.primaryLoader = primaryLoader
+        self.fallbackLoader = fallbackLoader
+    }
+
+    public func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        let task = TaskWrapper()
+        task.wrapper = primaryLoader.loadImageData(from: url) { [weak self] result in
+            switch result {
+                case .success:
+                    completion(result)
+
+                case .failure:
+                    task.wrapper = self?.fallbackLoader.loadImageData(from: url,
+                                                                      completion: completion)
+            }
+        }
+        return task
+    }
+}
